teams: Method to parse parents to teams which do not exist yet
This commit introduces a quick-and-dirty O(3n^2) mechanaic for
solving teams which have parents which do not exist yet on Github
but do in the teams folder.  Once parents are discovered, we can
now process team creation in order.  This can be done by the
reference to team object which is passed to the child object.

A `Sync` method is called on all team objects.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>