cmd: Introduce high-level `governctl` entrypoint program
This new entrypoint program acts as the basis for introducign sub-
commands (and additional functionality) to to the Governance
repository.  With it, we can streamline using the GitHub API for
all future programs.

For now, this commit introduces a nothing other than a simple
interface to for eventually connecting future sub-commands.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>