docs: Reduce copyright notice
This commit essentialy squashes the copyright notice such that it
references a well-known SPDX identifier which can be used to reference
the license.  Additionally a new `LICENSE.md` file is introduced into
the repository.  The boilerplate text that is used at the start of each
file adds unnecessary size text to each document/file.

Signed-off-by: Alexander Jung <alex@unikraft.io>