fix(governctl): Add unknown commits as exception
Because of how governctl structures temp repos it makes it
impossible for checkpatch to look behind and check if a
commit in a body is referenced correctly. This is fine.
It can be checked by hand with the script and can also
be observed in GitHub if the sha mentioned is invalid.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>